services-preview/api-service/requirements.txt
# api-service 依赖
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
psycopg[binary,pool]>=3.1
python-dotenv>=1.0.0
